user section redesign logs
