+++ src/components/admin/index.admincomp.tsx
@@ -15,3 +15,13 @@ export * from './view/Header';
 export * from './sections/SeccionMascota';
 
 export * from './sections/UserSection';
+
+export * from './usuarios/formularios/CrearPersonalForm';
+
+export * from './usuarios/listados/BitacoraList';
+
+export * from './usuarios/listados/ClienteList';
+
+export * from './usuarios/listados/PersonalList';
+
+export * from './usuarios/listados/UsuarioList';
